renamed PublicOrgSettings to OrgSettings
